majour fixing

+++ 0x00-hello_world/6-size.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 /**
  * main - A program that print different type sizes
- * Return: 0
+ * Return: 0(success)
  */
 int main(void)
 {
@@ -11,11 +11,11 @@ int main(void)
 	long long int d;
 	float f;
 
-printf("size of char: %lu bytes(s)\n", (unsigned long)sizeof(a));
-printf("size of int: %lu bytes(s)\n", (unsigned long)sizeof(b));
-printf("size of long int: %lu bytes(s)\n", (unsigned long)sizeof(c));
-printf("size of long logn int: %lu bytes(s)\n", (unsigned long)sizeof(d));
-printf("size of float: %lu bytes(s)\n", (unsigned long)sizeof(f));
+printf("size of char: %lu bytes(s)\n", sizeof(a));
+printf("size of int: %lu bytes(s)\n", sizeof(b));
+printf("size of long int: %lu bytes(s)\n", sizeof(c));
+printf("size of long logn int: %lu bytes(s)\n", sizeof(d));
+printf("size of float: %lu bytes(s)\n", sizeof(f));
 return (0);
 }
